fix eventhandler on spigot servers
